[demultiplexer] moved the simple throttler implementation in an aggregation internal util package.
